[FE] fix:axios interceptor 오류 수정
