+++ client/src/Pages/Domain.tsx
@@ -11,12 +11,11 @@ import Post from './Post';
 import PostEdit from './PostEdit';
 import AddMarker from './AddMarker';
 import NotFound from './NotFound';
-import { axiosGetRefresh } from '../util/UserApi';
 import PrivateRoute from '../Components/PrivateRouter';
-
-axiosGetRefresh;
+import { axiosRefresh } from '../util/Inter';
 
 export default function Router() {
+  axiosRefresh;
   return (
     <BrowserRouter>
       <Routes>
